refactor: generate test class more detail
